requirements.txt
asgiref==3.8.1
Django==5.0.7
sqlparse==0.5.1
pillow==10.3.0
djangorestframework~=3.15.2
freezegun==1.5.1
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==8.2.1
pytest-django==4.8.0
pytest-mock==3.14.0
